Separate a base class to test commands
