+++ Tests/Functional/CommandTest.php
@@ -4,55 +4,12 @@ namespace ShonM\ResqueBundle\Tests\Functional;
 
 use Symfony\Component\Console\Output\Output;
 use Symfony\Component\Console\Input\StringInput;
-use Symfony\Bundle\FrameworkBundle\Console\Application;
 
 use Resque\Event;
 use Resque\Job\Status;
 
-class BaseTest extends \PHPUnit_Framework_TestCase
+class CommandTest extends BaseCommandTest
 {
-    private $app;
-    private $kernel;
-    private static $container;
-    private $resque;
-
-    protected function setUp()
-    {
-        $this->kernel = new AppKernel();
-
-        $this->app = new Application($this->kernel);
-        $this->app->setAutoExit(false);
-        $this->app->setCatchExceptions(false);
-
-        $this->resque = $this->getContainer()->get('resque');
-    }
-
-    protected function tearDown()
-    {
-        $this->resque = null;
-        Event::clearListeners();
-    }
-
-    private function getContainer()
-    {
-        if (! self::$container) {
-            $this->kernel->boot();
-            self::$container = $this->kernel->getContainer();
-        }
-
-        return self::$container;
-    }
-
-    private function runCommand($command)
-    {
-        $output = new MemoryOutput();
-        $input = new StringInput($command);
-
-        $this->app->run($input, $output);
-
-        return $output->getOutput();
-    }
-
     /**
      * @expectedException \RuntimeException
      */
@@ -63,7 +20,7 @@ class BaseTest extends \PHPUnit_Framework_TestCase
 
     public function testSuccessfulEnqueue()
     {
-        $jobId = $this->resque->add('ShonM\ResqueBundle\Job\TestJob', 'test');
+        $jobId = $this->getResque()->add('ShonM\ResqueBundle\Job\TestJob', 'test');
 
         $this->assertTrue(is_string($jobId));
 
@@ -105,7 +62,7 @@ class BaseTest extends \PHPUnit_Framework_TestCase
      */
     public function testQueueSize()
     {
-        $size = $this->resque->size('test');
+        $size = $this->getResque()->size('test');
 
         $this->assertEquals(1, $size);
     }
@@ -122,24 +79,5 @@ class BaseTest extends \PHPUnit_Framework_TestCase
         $status = $this->runCommand('resque:job:status ' . $job);
 
         $this->assertEquals(Status::STATUS_COMPLETE, $status);
-    }
-}
-
-class MemoryOutput extends Output
-{
-    private $output;
-
-    protected function doWrite($message, $newline)
-    {
-        $this->output .= $message;
-
-        if ($newline) {
-            $this->output .= "\n";
-        }
-    }
-
-    public function getOutput()
-    {
-        return $this->output;
     }
 }
